Split up into different install methods
